Update conversational-context.md
All of the recommended changes below are for the "Using context to enable Intents" part of the guide. 

Recommendation 1: The speech for def handle_yes_honey_intent
Since self.milk = False, milk must return False in the def handle_yes_honey_intent when one refuses Milk. I also tried it out in practice and the speech needed to be flipped so that everything lines up with the correct answer at the end. 

Recommendation 2: The single apostrophe in the speech
The single apostrophes in both the def handle_no_honey_intent and def handle_yes_honey_intent for the phrase "Here's" was causing problems in python. Nothing would work until I changed the "Here's" to "Heres" maintaining single apostrophe throughout. I thought it simpler just to remove the single apostrophe, but maybe the convention should be double apostrophes for the self.speak. 

Recommendation 3: The import library
I thought it would be useful to have the library imported for contexts in the guide. Simply invoking adds_context or removes_context without the import library did not work. I was able to find it however in the Mycroft-Core (I think), but that was difficult as a newbie.