update url
have fixed the original "/privacy-policy" url which currently redirects to "/embed-privacy-policy".
So have now removed "embed-" from the url here. 
If we change the location of that privacy policy in the future, where ever the /privacy-policy url points to should be the correct location, then only need to change one thing to update everywhere.