Clean up converse() code
The method signature was different from the parent class, and the `else:` block was redundant.

As an aside, I'm a new skill developer and this documentation has created more questions than answers for me.  From what I'm reading here, all I should need to do is add a `converse()` method and it'll be triggered before the intent handler... except in my experience it's not triggered at all.